refactor ~bacdive~ => microbe
