+++ .github/workflows/deploy-docs.yml
@@ -28,11 +28,11 @@ jobs:
         
     - name: Build documentation.
       run: |
-        echo ${{ secrets.GH_TOKEN }} >> kg_bacdive/token.txt
+        echo ${{ secrets.GH_TOKEN }} >> kg_microbe/token.txt
         mkdir gh-pages
         touch gh-pages/.nojekyll
         cd docs/
-        poetry run sphinx-apidoc -o . ../kg_bacdive/ --ext-autodoc -f
+        poetry run sphinx-apidoc -o . ../kg_microbe/ --ext-autodoc -f
         poetry run sphinx-build -b html . _build
         cp -r _build/* ../gh-pages/
