+++ .github/workflows/qc.yml
@@ -1,4 +1,4 @@
-name: kg-bacdive QC
+name: kg-microbe QC
 
 on:
   push:
