+++ docs/conf.py
@@ -7,7 +7,7 @@ import os
 import re
 import sys
 from datetime import date
-from kg_bacdive import __version__
+from kg_microbe import __version__
 # -- Project information -----------------------------------------------------
 # https://www.sphinx-doc.org/en/master/usage/configuration.html#project-information
