+++ kg_microbe/__init__.py
@@ -1,4 +1,4 @@
-"""kg-bacdive package."""
+"""kg-microbe package."""
 from importlib import metadata
 
 from .download import download
