+++ kg_microbe/transform_utils/example_transform/example_transform.py
@@ -20,7 +20,7 @@ from typing import Optional, Union
 
 from transform_utils.transform import Transform
 
-from kg_bacdive.utils.robot_utils import convert_to_json, extract_convert_to_json
+from kg_microbe.utils.robot_utils import convert_to_json, extract_convert_to_json
 
 
 class YourTransform(Transform):
