+++ kg_microbe/transform_utils/traits/metabolism_map.yaml
