+++ merge.yaml
@@ -12,7 +12,7 @@ configuration:
     # define the type for non-canonical properties for RDF export
 
 merged_graph:
-  name: kg-bacdive graph
+  name: kg-microbe graph
   source:
     ncbitaxon:
       name: "NCBITaxon"
@@ -52,7 +52,7 @@ merged_graph:
   operations:
     - name: kgx.graph_operations.summarize_graph.generate_graph_stats
       args:
-        graph_name: kg-bacdive graph
+        graph_name: kg-microbe graph
         filename: merged_graph_stats.yaml
         node_facet_properties:
           - provided_by
@@ -67,4 +67,4 @@ merged_graph:
     # merged-kg-nt:
     #   format: nt
     #   compression: gz
-    #   filename: kg_bacdive.nt.gz
+    #   filename: kg_microbe.nt.gz
