+++ templates/README.toplevel
@@ -1,4 +1,4 @@
-The subdirectories here each contain a build of the kg-bacdive knowledge graph.
+The subdirectories here each contain a build of the kg-microbe knowledge graph.
 The directories are named with a date (ISO 8601 format, YYYYMMDD) according to when
 the build was produced. The current/ directory is the most recent build.
