+++ tox.ini
@@ -38,8 +38,8 @@ deps =
     ruff
 skip_install = true
 commands =
-    black kg_bacdive/ tests/
-    ruff --fix kg_bacdive/ tests/
+    black kg_microbe/ tests/
+    ruff --fix kg_microbe/ tests/
 description = Run linters.
 
 # This is used for QC checks.
@@ -49,8 +49,8 @@ deps =
     ruff
 skip_install = true
 commands =
-    black --check --diff kg_bacdive/ tests/
-    ruff check kg_bacdive/ tests/
+    black --check --diff kg_microbe/ tests/
+    ruff check kg_microbe/ tests/
 description = Run linters.
 
 [testenv:doclint]
@@ -67,7 +67,7 @@ skip_install = true
 deps = 
     codespell
     tomli  # required for getting config from pyproject.toml
-commands = codespell kg_bacdive/ tests/ -S kg_bacdive/transform_utils/*/tmp/*
+commands = codespell kg_microbe/ tests/ -S kg_microbe/transform_utils/*/tmp/*
 
 [testenv:codespell-write]
 description = Run spell checker and write corrections.
@@ -75,12 +75,12 @@ skip_install = true
 deps = 
     codespell
     tomli
-commands = codespell kg_bacdive/ tests/ --write-changes -S kg_bacdive/transform_utils/*/tmp/*
+commands = codespell kg_microbe/ tests/ --write-changes -S kg_microbe/transform_utils/*/tmp/*
 
 [testenv:docstr-coverage]
 skip_install = true
 deps =
     docstr-coverage
 commands =
-    docstr-coverage kg_bacdive/ tests/ --skip-private --skip-magic
+    docstr-coverage kg_microbe/ tests/ --skip-private --skip-magic
 description = Run the docstr-coverage tool to check documentation coverage
